Add .gitignore files and Cargo.lock dependency file
This commit includes the addition of .gitignore files in both the main project directory and subdirectory nostr_db. These help specify intentionally untracked files that git should ignore. This commit also adds the Cargo.lock file to the project, which contains the exact versions of the dependencies needed for the project to build and run correctly. This file was generated after building the project using Cargo, Rust's package manager.